docs: fix quoting issue, clarify naming in auth.md
The steps for creating serving certificates were missing a pair of
quotes, which could have lead to some confusion.

This also clarifies the naming of the serving certs a bit: many places
refer to the cert and key as `.crt` and `.key`, but `cfssl` outputs them
as `.pem` and `-key.pem`.