Do not initialize custom REST handler during ApiVersion initialization
ApiVersion variable from the versioned package is initialized both
in the apiserver and in the controller-manager because they both
import the versioned package.
When a custom REST handler is defined, its NewFooREST() constructor is
being called directly during ApiVersion initialization. This constructor
may contain some other logic besides simply initializing the custom REST
registry type. For example it may establish a connection to some service.
And we'll end up with an extra unused connection to the external
service, established from the controller manager.

Passing NewFooREST as a function value to builders.NewApiResourceWithStorage()
and then calling it from versionedResourceBuilder.registerEndpoints()
solves this problem.

Fixes #92.